requirements.txt
# Core Framework
langchain==0.1.0
langgraph==0.0.20
langchain-community==0.0.10
langfuse==2.0.0

# LLM Providers
ollama==0.1.0
openai==1.0.0
groq==0.4.0

# Vector Store & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Web Tools
tavily-python==0.3.0
requests==2.31.0
beautifulsoup4==4.12.0

# Web Framework
streamlit==1.29.0
gradio==4.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23

# Monitoring
langsmith==0.0.70

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Logging
loguru==0.7.2